代码结构修改
