+++ Flutter/lib/engine/dokit_binding.dart
@@ -3,6 +3,7 @@ import 'dart:convert';
 import 'dart:ui';
 import 'package:dokit/kit/apm/apm.dart';
 import 'package:dokit/kit/apm/method_channel_kit.dart';
+import 'package:dokit/kit/kit.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
