+++ Flutter/lib/kit/apm/apm.dart
@@ -1,4 +1,3 @@
-import 'dart:collection';
 
 import 'package:dokit/kit/apm/fps_kit.dart';
 import 'package:dokit/kit/apm/log_kit.dart';
@@ -8,6 +7,7 @@ import 'package:dokit/kit/apm/route_kit.dart';
 import 'package:dokit/ui/resident_page.dart';
 import 'package:flutter/material.dart';
 
+import '../kit.dart';
 import 'http_kit.dart';
 
 class ApmKitManager {
@@ -44,52 +44,6 @@ class ApmKitManager {
     kitMap.forEach((key, kit) {
       kit.start();
     });
-  }
-}
-
-abstract class IInfo {
-  dynamic getValue();
-}
-
-abstract class IStorage {
-  bool save(IInfo info);
-
-  bool contains(IInfo info);
-
-  List<IInfo> getAll();
-}
-
-abstract class IKit {
-  String getKitName();
-
-  String getIcon();
-
-  void tabAction();
-}
-
-class CommonStorage implements IStorage {
-  final int maxCount;
-  Queue<IInfo> items = new Queue();
-
-  CommonStorage({this.maxCount = 100});
-
-  @override
-  List<IInfo> getAll() {
-    return items.toList();
-  }
-
-  @override
-  bool save(IInfo info) {
-    if (items.length >= maxCount) {
-      items.removeFirst();
-    }
-    items.add(info);
-    return true;
-  }
-
-  @override
-  bool contains(IInfo info) {
-    return items.contains(info);
   }
 }
