+++ Flutter/lib/kit/visual/view_check.dart
@@ -1,5 +1,4 @@
 import 'package:dokit/dokit.dart';
-import 'package:dokit/kit/apm/apm.dart';
 import 'package:dokit/kit/visual/visual.dart';
 import 'package:dokit/ui/dokit_app.dart';
 import 'package:dokit/ui/dokit_btn.dart';
