+++ Flutter/lib/widget/fps_chart.dart
@@ -1,6 +1,6 @@
 import 'dart:math';
 
-import 'package:dokit/kit/apm/apm.dart';
+import 'package:dokit/kit/kit.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
